pyproject.toml
[tool.poetry]
name = "django-tsm"
version = "0.1.0"
description = ""
authors = ["ZiG <x6vital@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.2"
djangorestframework = "^3.15.2"
markdown = "^3.7"
django-filter = "^24.3"
drf-yasg = "^1.21.7"
celery = "^5.4.0"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
redis = "^5.1.1"
django-celery-results = "^2.5.1"
flower = "^2.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
